move dev to workspace
