+++ scaleph-ui-react/src/pages/Project/Workspace/Cluster/ConfigOptions/components/K8sResource.tsx
