+++ scaleph-ui-react/src/pages/Project/Workspace/Cluster/ConfigOptions/components/Resource.tsx
