--- scaleph-ui-react/src/pages/Project/Workspace/Cluster/index.tsx
@@ -1,4 +1,0 @@
-const ClusterConfigView: React.FC = () => {
-  return <>cluster config</>;
-};
-export default ClusterConfigView;
